+++ service/flex/v1/plugins/api_op_client.go
@@ -1,0 +1,17 @@
+// Package plugins contains auto-generated files. DO NOT MODIFY
+package plugins
+
+import "github.com/RJPearson94/twilio-sdk-go/client"
+
+// Client for managing plugin resources
+// See https://www.twilio.com/docs/flex/developer/plugins/api/plugin for more details
+type Client struct {
+	client *client.Client
+}
+
+// New creates a new instance of the plugins client
+func New(client *client.Client) *Client {
+	return &Client{
+		client: client,
+	}
+}